+++ lib/components/home/index.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class Home extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+      void _incrementCounter() {
+          Navigator.pushNamed(context, '/class');
+      }
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Home"),
+      ),
+      body: Center(
+        child: 
+          Text(
+            'HelloWorld',
+          ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: _incrementCounter,
+        tooltip: 'Increment',
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
